refactor: move user login to 'login', move admin login to 'login_admin'.
